+++ changelog.md
@@ -13,6 +13,7 @@ Additions
 - Salepoint: Similar to the vendor, but allows players to queue a transaction made up of multiple purchases through a Portable Item/Fluid/Energy Interface
 - Add ponders for all shop blocks
 - Brass Depositors with a price of zero can accept coins from item transfer (hoppers, chutes, etc.)
+- Bank Cards and Authorized Cards can be held in the offhand to pay at Create 6 tablecloth shops
 
 Fixes
 - Fix coins stacked above 127 visually disappearing on the client in Brass Depositors